# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "title": "Electronic notifications",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/",
    "paths": {
...
        "/api/recipients": {
            "get": {
                "description": "Возвращает всех доуступных получателей с опциональной фильтрацией по ФИО",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Получатели"
                ],
                "summary": "Получить всех получателей",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ФИО для фильтрации",
                        "name": "fio",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemes.GetAllRecipientsResponse"
                        }
                    }
                }
            }
        },
        "/api/recipients/": {
            "post": {
                "description": "Добавить нового получателя",
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "Получатели"
                ],
                "summary": "Добавить получателя",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Изображение получателя",
                        "name": "image",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "ФИО",
                        "name": "fio",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Почта",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Возраст",
                        "name": "age",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Адрес",
                        "name": "adress",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/recipients/{recipient_id}": {
            "get": {
                "description": "Возвращает более подробную информацию об одном получателе",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Получатели"
                ],
                "summary": "Получить одного получателя",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id получателя",
                        "name": "recipient_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.Recipient"
                        }
                    }
                }
            },
            "put": {
                "description": "Изменить данные полей о получателе",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Получатели"
                ],
                "summary": "Изменить получателя",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор получателя",
                        "name": "recipient_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ФИО",
                        "name": "fio",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Почта",
                        "name": "email",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "Возраст",
                        "name": "age",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Изображение получателя",
                        "name": "image",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Адрес",
                        "name": "adress",
                        "in": "formData"
                    }
                ],
                "responses": {}
            },
            "delete": {
                "description": "Удаляет получателя по id",
                "tags": [
                    "Получатели"
                ],
                "summary": "Удалить получателя",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id получателя",
                        "name": "recipient_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/recipients/{recipient_id}/add_to_notification": {
            "post": {
                "description": "Добавить выбранного получателя в черновик уведомления",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Получатели"
                ],
                "summary": "Добавить в уведомление",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id получателя",
                        "name": "recipient_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemes.AllRecipientsResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ds.Recipient": {
            "type": "object",
            "required": [
                "adress",
                "age",
                "email",
                "fio"
            ],
            "properties": {
                "adress": {
                    "type": "string"
                },
                "age": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                },
                "fio": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
        "schemes.AllRecipientsResponse": {
            "type": "object",
            "properties": {
                "recipients": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ds.Recipient"
                    }
                }
            }
        },
        "schemes.GetAllRecipientsResponse": {
            "type": "object",
            "properties": {
                "draft_notification": {
                    "$ref": "#/definitions/schemes.NotificationShort"
                },
                "recipients": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ds.Recipient"
                    }
                }
            }
        },
...
        "schemes.NotificationShort": {
            "type": "object",
            "properties": {
                "recipient_count": {
                    "type": "integer"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
...
    }
}